Backoff module extraction (#873)
