+++ backoff/src/main/java/io/split/android/client/backoff/FixedIntervalBackoffCounter.java
@@ -1,4 +1,4 @@
-package io.split.android.client.service.sseclient;
+package io.split.android.client.backoff;
 
 public class FixedIntervalBackoffCounter implements BackoffCounter {
