+++ main/build.gradle
@@ -54,6 +54,7 @@ dependencies {
     api clientModuleProject('api')
     api clientModuleProject('http-api')
     api clientModuleProject('fallback')
+    api clientModuleProject('backoff')
     // Internal module dependencies
     implementation clientModuleProject('http')
     implementation clientModuleProject('events-domain')
