+++ main/src/test/java/io/split/android/client/service/MySegmentsSyncTaskTest.java
@@ -40,7 +40,7 @@ import io.split.android.client.service.http.HttpFetcher;
 import io.split.android.client.service.http.HttpFetcherException;
 import io.split.android.client.service.mysegments.MySegmentsSyncTask;
 import io.split.android.client.service.mysegments.MySegmentsSyncTaskConfig;
-import io.split.android.client.service.sseclient.BackoffCounter;
+import io.split.android.client.backoff.BackoffCounter;
 import io.split.android.client.service.synchronizer.MySegmentsChangeChecker;
 import io.split.android.client.storage.mysegments.MySegmentsStorage;
 import io.split.android.client.telemetry.model.OperationType;
