+++ settings.gradle
@@ -8,3 +8,4 @@ include ':fallback'
 include ':main'
 include ':events'
 include ':events-domain'
+include ':backoff'
